+++ src/js/page/index.js
@@ -1,7 +1,7 @@
 import JSZip from 'jszip'
-import Papa from 'papaparse'
 
 import { Helper } from '../Helper.js'
+import { Csv } from '../Csv.js'
 import { Dom } from '../Dom.js'
 import { Filter } from '../Filter.js'
 import { Build } from '../Build.js'
@@ -9,135 +9,6 @@ import { Build } from '../Build.js'
 import '../../css/style.css'
 import '../../css/dark.css'
 import '../../images/meta-image.png'
-
-/**
- * Search the zip file for the required CSV file
- */
-function findCsvFile (zip) {
-  const regex = /\.csv$/
-  let filename = null
-
-  for (const [key, info] of Object.entries(zip.files)) {
-    if (regex.test(info.name) === true) {
-      console.log(`CSV file: ${info.name} (${key})`)
-      filename = info.name
-    }
-  }
-
-  if (filename === null) {
-    throw Error('No CSV file not found in zip file.')
-  }
-
-  return filename
-}
-
-/**
- * Process the CSV file and create an array of tweets
- */
-async function processCsvFile (filename, zip) {
-  return new Promise((resolve) => {
-    zip.file(filename).async('text').then(function (content) {
-      const results = Papa.parse(content)
-
-      const data = {
-        tweets: [],
-        users: [],
-        stats: {
-          users: 0,
-          tweets: 0,
-          images: 0,
-          videos: 0
-        }
-      }
-
-      let rowsAfter = 0
-      for (let index = 0; index < results.data.length; index++) {
-        if (results.data[index][0] === 'Tweet date') {
-          rowsAfter = index
-          break
-        }
-      }
-
-      results.data.forEach((row, index) => {
-        if (index > rowsAfter) {
-          if (Helper.getUserIndex(row[3], data.users) === -1) {
-            const user = {
-              display_name: row[2],
-              username: row[3],
-              tweets: 0
-            }
-
-            data.stats.users++
-            data.users.push(user)
-          }
-
-          const id = Helper.getIdFromUrl(row[4])
-          const tweetIndex = Helper.getTweetIndex(id, data.tweets)
-          const userIndex = Helper.getUserIndex(row[3], data.users)
-
-          if (tweetIndex === -1) {
-            const tweet = {
-              date: row[0],
-              display_name: row[2],
-              username: row[3],
-              url: row[4],
-              id,
-              media: [],
-              stats: {
-                images: 0,
-                videos: 0,
-                replies: Number(row[10]),
-                retweets: Number(row[11]),
-                likes: Number(row[12])
-              },
-              remarks: row[8],
-              text: row[9]
-            }
-
-            const media = {}
-            if (row[5] !== 'No media') {
-              media.type = row[5]
-              media.url = row[6]
-              media.filename = row[7]
-              tweet.media.push(media)
-
-              if (row[5] === 'Image' || row[5] === 'GIF') {
-                data.stats.images++
-                tweet.stats.images++
-              } else {
-                data.stats.videos++
-                tweet.stats.videos++
-              }
-            }
-
-            data.stats.tweets++
-            data.users[userIndex].tweets++
-
-            data.tweets.push(tweet)
-          } else if (tweetIndex !== -1 && row[5] !== 'No media') {
-            const media = {
-              type: row[5],
-              url: row[6],
-              filename: row[7]
-            }
-
-            if (row[5] === 'Image' || row[5] === 'GIF') {
-              data.stats.images++
-              data.tweets[tweetIndex].stats.images++
-            } else {
-              data.stats.videos++
-              data.tweets[tweetIndex].stats.videos++
-            }
-
-            data.tweets[tweetIndex].media.push(media)
-          }
-        }
-      })
-
-      resolve(data)
-    })
-  })
-}
 
 /**
  * Load a zip archive and display tweets
@@ -160,8 +31,8 @@ function loadFile (input) {
   reader.onload = function (ev) {
     JSZip.loadAsync(ev.target.result)
       .then(async function (zip) {
-        const csvFilename = findCsvFile(zip)
-        const data = await processCsvFile(csvFilename, zip)
+        const csvFilename = Csv.find(zip)
+        const data = await Csv.process(csvFilename, zip)
         const autoload = document.getElementById('autoload').checked
 
         Dom.hide('loading')
